+++ container/docker/templates/conductor-dockerfile.j2
