+++ setup.py
@@ -31,7 +31,7 @@ if container.ENV == 'host':
     setup_kwargs = dict(
         install_requires=[str(ir.req) for ir in install_reqs],
         tests_require=[
-            'ansible>=2.3.0',
+            'ansible==2.4.0',
             'pytest>=3',
             'docker>=2.1'
         ],
@@ -39,7 +39,7 @@ if container.ENV == 'host':
             'docker': ['docker>=2.1'],
         },
         dependency_links=[
-            'git+https://github.com/ansible/ansible@develop#egg=ansible'
+            'git+https://github.com/ansible/ansible@devel#egg=ansible-2.4.0'
         ],
         cmdclass={'test': PlaybookAsTests},
         entry_points={
